RaspberryPi/requirements.txt
boto3
requests
python-dotenv
pip3 install RPi.GPIO